modmail_bot/requirements.txt
discord.py>=2.3.2
motor>=3.3.2
pymongo>=4.6.1
python-dotenv>=1.0.0
aiofiles>=23.2.1
Pillow>=10.0.0
